+++ packages/ripple-ui-core/src/components/search-bar/RplSearchBar.vue
@@ -131,6 +131,12 @@ const handleSubmit = (type: 'button' | 'enter') => {
 }
 
 const handleInputChange = (e) => {
+  // It was requested in SD-1817 that the search bar immediately "submits" if a user clears the field through any means (backspace etc.)
+  if (props.submitOnClear && e.target.value === '') {
+    handleClear()
+    return
+  }
+
   internalValue.value = e.target.value
   emit('update:inputValue', e.target.value)
   isOpen.value = true