+++ packages/ripple-ui-core/vite.config.ts
@@ -1,7 +1,6 @@
 import { defineConfig } from 'vite'
 import path from 'node:path'
 import vue from '@vitejs/plugin-vue'
-import dts from 'vite-plugin-dts'
 import vitePlugins from './src/vite.plugins'
 
 // https://vitejs.dev/config/
@@ -13,7 +12,7 @@ export default defineConfig({
       vue: 'vue/dist/vue.esm-bundler.js'
     }
   },
-  plugins: [vue(), dts()].concat(vitePlugins),
+  plugins: [vue()].concat(vitePlugins),
   build: {
     emptyOutDir: false,
     lib: {